% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdat.R
\name{getdat}
\alias{getdat}
\title{getdat}
\usage{
getdat(n = 1000, xylim = 1000)
}
\arguments{
\item{n}{Number of edges}

\item{xylim}{Maximal x and y values}
}
\value{
A \code{data.frame} of randomly wandering edges
}
\description{
Make a single trail from a bunch of random connected edges
}
