% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdat.R
\name{getpoints}
\alias{getpoints}
\title{getpoints}
\usage{
getpoints(n = 1000, xylim = 1000)
}
\arguments{
\item{n}{Number of points}

\item{xylim}{Maximal x and y values}
}
\value{
A \code{data.frame} of random points
}
\description{
Get a bunch of random points along with colour attributes
}
