% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgplot.R
\name{svgplot}
\alias{svgplot}
\title{svgplot}
\usage{
svgplot(dat, filename)
}
\arguments{
\item{dat}{A test \code{data.frame} from \link{getdat}}

\item{filename}{Name of \code{html} file to write \code{svg} data}
}
\value{
Nothing
}
\description{
Plot data as \code{html}-formatted \code{svg} file.
}
