% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgplot.R
\name{svgplot_lines}
\alias{svgplot_lines}
\title{svgplot_lines}
\usage{
svgplot_lines(dat, filename, html = TRUE)
}
\arguments{
\item{dat}{A test \code{data.frame} from \link{getlines}}

\item{filename}{Name of \code{html} file to write \code{svg} data}

\item{html}{If \code{TRUE}, produce file in \code{html} format; otherwise
straight \code{svg}.}
}
\value{
Nothing
}
\description{
Plot line data as \code{html}-formatted \code{svg} file.
}
